Vectorize input checks, add isfinite(), fix col output for row input
Previously, the input checking was done using commands like
any(~isreal(thisGeometricHeight_km)), which returned a vector when
thisGeometricHeight_km was a 2-D matrix, and so on.

It has been replaced with any(~isreal(thisGeometricHeight_km,'all'))
which always returns a scalar for multi-dimensional array input.

Also introduced isfinite to check if the input heights are NaN/Inf
anywhere.

Further, when input was a row vector, the outputs were column vectors
earlier. That has been fixed by intitializing the outputs with zeros
of the same size as that of the input.