[FinGeometry] Return property names also after input parsing
If only values are returned from parseInputs, they may not be assigned to
the correct property in the constructor, because the property names are
read from the defaults file, and their ordering cannot be guaranteed (it
could, but that's inconvenient for the user to never change the order).

So, the names of the properties corresponding to each parsed value are
also returned from parseInputs.