Modify defaultInputs to read arrays as property values, and return struct
If the default value of a property is an array, it can now be read from
the file. Further, the output of the function is a struct with fieldnames
equal to that of the names in the file.

The defaults file for FinGeometry has been modified accordingly, and
a new one has been created for FoilGeometry.